Add files via upload
The k in k-nearest neighbours signifies that instead of using only the closest neighbour to the new data point, we can
consider any fixed number of k of neighbours in training (for example, the closest three or five neighbours). Then, we can
make a prediction using the majority class among these neighbours. The KNN object encapsulates the algorithm that will
be used to build the model from the training data, as well as the algorithm to make predictions on new data points. It
will also hold the information that the algorithm has extracted from the training data. In the case of
KNeighborsClassifier, it will just store the training set